LayoutTests/imported/w3c/web-platform-tests/html/browsers/windows/nested-browsing-contexts/window-top-null-expected.txt


PASS `window.top` is null when browsing context container element removed
PASS `window.top`null when any ancestor browsing context container removed
